requirements.txt
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
mediapipe>=0.10.0
PyQt5>=5.15.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
websocket-client>=1.6.0
pynput>=1.7.0
requests>=2.31.0
python-dotenv>=1.0.0
Flask>=2.3.0
Flask-CORS>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.9.0
python-engineio>=4.7.0
